RF: Move unique metadata value curation from extraction to aggregation
